refactor: extract logging sanitisation into dedicated service
